refactor: update zod schema
