+++ src/controller/calledController/index.ts
@@ -1,4 +1,4 @@
+import { CalledController } from "./calledController";
 import { calledUseCase } from "../../usecases/calledUseCases";
-import { CalledController } from "./called.controller";
 
 export const calledController = new CalledController(calledUseCase)
